+++ orttraining/orttraining/python/training/ortmodule/__init__.py
@@ -120,6 +120,7 @@ def _are_deterministic_algorithms_enabled():
     return ORTMODULE_IS_DETERMINISTIC
 
 
+from .graph_transformer_registry import register_graph_transformer  # noqa: E402, F401
 from .options import DebugOptions, LogLevel  # noqa: E402, F401
 
 # ORTModule must be loaded only after all validation passes
